Added tribunal & fixes
Tribunal is now working for the users part. Which means user will create
a case (an ongoing which need to be reviewed) and 3 others users will be
randomly selected among the user list (excluding the two persons
involved in the case).
Those persons now have access to the tribunal and can either deny or
confirm the case.

- fixed missing font-awesome files
- Collections updates